# specs/002-api-contract/openapi-v0.yaml
openapi: 3.1.0
info:
  title: ERP-ALL API
  version: 0.1.0-draft
  description: >
    契约草案 v0（EA-002）。五个核心上下文：identity / channel / catalog / listing / order + 采购门户。
    全局约定见同目录 README.md；数据语义以 specs/001-domain-model/ 为准。
    x-permission = 所需权限点（permission.code），网关中间件按此校验。
servers:
  - url: /api/v1
    description: 内部 API（JWT audience=erp）
  - url: /portal/v1
    description: 采购门户 API（JWT audience=portal，仅 Portal tag 下端点）
...
security:
  - bearerAuth: []
paths:
...
  /gtin-pool:
    get:
      tags: [Catalog]
      x-permission: catalog.gtin_read
      parameters:
        - {$ref: '#/components/parameters/page'}
        - {$ref: '#/components/parameters/size'}
        - {name: status, in: query, schema: {type: string}}
        - {name: gtin_kind, in: query, schema: {type: string}}
      responses: {'200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/GtinPage'}}}}}
...
  /listings:
    get:
      tags: [Listing]
      x-permission: listing.read
      parameters:
        - {$ref: '#/components/parameters/page'}
        - {$ref: '#/components/parameters/size'}
        - {name: store_id, in: query, schema: {type: integer}}
        - {name: status, in: query, schema: {type: string}}
        - {name: offer_mode, in: query, schema: {type: string, enum: [build, match]}}
        - {name: error_code, in: query, schema: {type: string}}
      responses: {'200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/ListingPage'}}}}}
  /listings/allocate:
    post:
      tags: [Listing]
      summary: 分配上架（批量建 draft：产品×店铺×模式；含去重/品牌占用/GTIN 预占）
      x-permission: listing.allocate
      parameters: [{$ref: '#/components/parameters/idempotencyKey'}]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [product_ids, store_id, offer_mode]
              properties:
                product_ids: {type: array, items: {type: integer}, maxItems: 500}
                store_id: {type: integer}
                offer_mode: {type: string, enum: [build, match]}
      responses:
        '200':
          description: 逐品结果（部分成功正常返回）
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: {type: array, items: {$ref: '#/components/schemas/Listing'}}
                  rejected:
                    type: array
                    items:
                      type: object
                      properties:
                        product_id: {type: integer}
                        code: {type: string, examples: [LISTING_DUP_IN_TEAM, BRAND_OCCUPIED_OTHER_STORE, GTIN_POOL_EMPTY, PRODUCT_NOT_READY]}
                        message: {type: string}
        '409': {description: 幂等键冲突（同键异载荷/处理中）, $ref: '#/components/responses/Error'}
  /listings/{listingId}:
    get:
      tags: [Listing]
      x-permission: listing.read
      parameters: [{$ref: '#/components/parameters/listingId'}]
      responses: {'200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/ListingDetail'}}}}}
  /listings/submit:
    post:
      tags: [Listing]
      summary: 批量提交上架（异步：构建 spec→组 feed→网关提交；消耗 listing_create 配额）
      x-permission: listing.submit
      parameters: [{$ref: '#/components/parameters/idempotencyKey'}]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [listing_ids]
              properties: {listing_ids: {type: array, items: {type: integer}, maxItems: 500}}
      responses:
        '202': {description: 已入队, content: {application/json: {schema: {type: object, properties: {queued: {type: integer}, skipped: {type: array, items: {type: object}}}}}}}
        '409': {description: 幂等键冲突（同键异载荷/处理中）, $ref: '#/components/responses/Error'}
        '422': {description: 配额不足等业务拒绝, $ref: '#/components/responses/Error'}
  /listings/{listingId}/delist:
    post:
      tags: [Listing]
      summary: 下架（异步；消耗 listing_delete 配额）
      x-permission: listing.delist
      parameters: [{$ref: '#/components/parameters/listingId'}, {$ref: '#/components/parameters/idempotencyKey'}]
      responses: {'202': {description: 已入队}, '409': {description: 幂等键冲突（同键异载荷/处理中）, $ref: '#/components/responses/Error'}}
  /listings/{listingId}/retry:
    post:
      tags: [Listing]
      summary: failed 重投（按 error catalog 处置校验）
      x-permission: listing.submit
      parameters: [{$ref: '#/components/parameters/listingId'}]
      responses: {'202': {description: 已入队}, '409': {$ref: '#/components/responses/Error'}}
  /feeds:
    get:
      tags: [Listing]
      x-permission: listing.read
      parameters:
        - {$ref: '#/components/parameters/page'}
        - {$ref: '#/components/parameters/size'}
        - {name: store_id, in: query, schema: {type: integer}}
        - {name: feed_kind, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {type: string}}
      responses: {'200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/FeedPage'}}}}}
  /feeds/{feedId}:
    get:
      tags: [Listing]
      x-permission: listing.read
      parameters: [{name: feedId, in: path, required: true, schema: {type: integer}}]
      responses: {'200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/Feed'}}}}}
  /feeds/{feedId}/items:
    get:
      tags: [Listing]
      summary: item 级权威结果（headline 不可信）
      x-permission: listing.read
      parameters:
        - {name: feedId, in: path, required: true, schema: {type: integer}}
        - {$ref: '#/components/parameters/page'}
        - {$ref: '#/components/parameters/size'}
        - {name: status, in: query, schema: {type: string, enum: [pending, success, error]}}
      responses: {'200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/FeedItemPage'}}}}}
  /feeds/{feedId}/poll:
    post:
      tags: [Listing]
      summary: 手动轮询 feed 状态（item 级权威回写；beat 自动轮询 R2）
      x-permission: listing.submit
      parameters: [{name: feedId, in: path, required: true, schema: {type: integer}}]
      responses: {'200': {description: OK}, '422': {$ref: '#/components/responses/Error'}}
  /feeds/{feedId}/verify-back:
    post:
      tags: [Listing]
      summary: verify_pending 对账归位（adopt/lost）——提交无响应永不盲重试
      x-permission: listing.submit
      parameters: [{name: feedId, in: path, required: true, schema: {type: integer}}]
      responses: {'200': {description: OK}, '422': {$ref: '#/components/responses/Error'}}
  /gtin-pool/import:
    post:
      tags: [Catalog]
      summary: GTIN 批量导入（格式+GS1校验位校验；重复逐条报告）
      x-permission: catalog.import_write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gtins]
              properties:
                gtins: {type: array, items: {type: string}, maxItems: 10000}
                source: {type: string, enum: [generator_import, feishu_import, purchased], default: generator_import}
      responses: {'201': {description: '{imported, rejected[]}'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT, description: 内部 JWT（audience=erp）}
    portalAuth: {type: http, scheme: bearer, bearerFormat: JWT, description: 门户 JWT（audience=portal）}
  parameters:
    page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    size: {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
    idempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, maxLength: 64}}
...
    listingId: {name: listingId, in: path, required: true, schema: {type: integer}}
...
  responses:
    Error:
      description: 统一错误信封
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            detail: {type: object}
            request_id: {type: string}
    PageMeta:
      type: object
      properties:
        total: {type: integer}
        page: {type: integer}
        size: {type: integer}
...
    GtinItem:
      type: object
      properties:
        id: {type: integer}
        gtin: {type: string}
        gtin_kind: {type: string}
        status: {type: string}
        source: {type: string}
        held_listing_id: {type: [integer, 'null']}
        used_listing_id: {type: [integer, 'null']}
    GtinPage:
      allOf:
        - {$ref: '#/components/schemas/PageMeta'}
        - {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/GtinItem'}}}}
...
    Listing:
      type: object
      properties:
        id: {type: integer}
        store_id: {type: integer}
        product_id: {type: integer}
        offer_mode: {type: string}
        channel_sku: {type: string}
        gtin: {type: [string, 'null']}
        status: {type: string}
        error_code: {type: [string, 'null']}
        is_locked: {type: boolean}
        current_price: {type: [number, 'null']}
        current_inventory: {type: integer}
        published_at: {type: [string, 'null'], format: date-time}
    ListingDetail:
      allOf:
        - {$ref: '#/components/schemas/Listing'}
        - type: object
          properties:
            wpid: {type: [string, 'null']}
            channel_item_id: {type: [string, 'null']}
            end_date: {type: string, format: date}
            state_history:
              type: array
              items:
                type: object
                properties:
                  from_status: {type: string}
                  to_status: {type: string}
                  reason_code: {type: [string, 'null']}
                  occurred_at: {type: string, format: date-time}
            price_history:
              type: array
              items:
                type: object
                properties:
                  old_price: {type: [number, 'null']}
                  new_price: {type: number}
                  reason: {type: string}
                  occurred_at: {type: string, format: date-time}
    ListingPage:
      allOf:
        - {$ref: '#/components/schemas/PageMeta'}
        - {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Listing'}}}}
    Feed:
      type: object
      properties:
        id: {type: integer}
        store_id: {type: integer}
        channel_feed_id: {type: [string, 'null']}
        feed_kind: {type: string}
        status: {type: string}
        item_count: {type: integer}
        headline: {type: [object, 'null'], description: 渠道汇总（不可信，展示用）}
        submitted_at: {type: [string, 'null'], format: date-time}
        completed_at: {type: [string, 'null'], format: date-time}
    FeedPage:
      allOf:
        - {$ref: '#/components/schemas/PageMeta'}
        - {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Feed'}}}}
    FeedItem:
      type: object
      properties:
        id: {type: integer}
        listing_id: {type: integer}
        channel_sku: {type: string}
        status: {type: string}
        error_code: {type: [string, 'null']}
        error_msg: {type: [string, 'null']}
    FeedItemPage:
      allOf:
        - {$ref: '#/components/schemas/PageMeta'}
        - {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/FeedItem'}}}}
...